Creating and querying a data frame
Scenario
Miss Louw is so excited about your proposal on how to keep track of stock for Bennie’s Books, that she has decided to create a loyalty programme for repeat customers. She has provided you with raw data that indicates the personal information of clients (bennies_books_loyalty_data.txt). Miss Louw wants you to create a unique customer ID for each customer on the list, along with adding it to the table and extracting the data from the table. This will include:

all female customers
all customers over the age of 50
all customers with fewer than 20 purchases.
Objective
Create a data frame in R containing the information that Miss Louw provided, and then use functions to extract data from the table.